fixed casting to float (i/10) != float if return type is float. has to be casted

+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/commands/Speed.java
@@ -44,6 +44,6 @@ public class Speed implements CommandExecutor {
     }
 
     private float parseSpeed(String arg) {
-        return Integer.parseInt(arg)/10;
+        return (float)Integer.parseInt(arg)/10;
     }
 }
